test(cy): make node test helpers available
