+++ cypress/fixtures/Preview.md
@@ -1,0 +1,23 @@
+## Runs a spec
+
+empty
+
+---
+
+empty
+
+## Preserves a link
+
+[link text](https://nextcloud.com)
+
+---
+
+[link text](https://nextcloud.com)
+
+## Preserves a preview 
+
+[link text](https://nextcloud.com (Preview))
+
+---
+
+[link text](https://nextcloud.com (Preview))
